FEM: command package, move command modules in one module in new command package
