+++ src/Mod/Fem/CMakeLists.txt
@@ -55,6 +55,16 @@ INSTALL(
 
 INSTALL(
     FILES
+        femcommand/__init__.py
+        femcommand/commands.py
+        femcommand/_CommandFemEquation.py
+        femcommand/manager.py
+    DESTINATION
+        Mod/Fem/femcommand
+)
+
+INSTALL(
+    FILES
         feminterface/__init__.py
         feminterface/convert2TetGen.py
         feminterface/importCcxDatResults.py
@@ -154,38 +164,9 @@ INSTALL(
 
 INSTALL(
     FILES
-        PyGui/FemCommands.py
         PyGui/FemSelectionObserver.py
         PyGui/FemSelectionWidgets.py
         PyGui/__init__.py
-        PyGui/_CommandFemSolverElmer.py
-        PyGui/_CommandFemEquation.py
-        PyGui/_CommandFemConstraintBodyHeatSource.py
-        PyGui/_CommandFemConstraintFlowVelocity.py
-        PyGui/_CommandFemConstraintInitialFlowVelocity.py
-        PyGui/_CommandFemConstraintElectrostaticPotential.py
-        PyGui/_CommandFemAnalysis.py
-        PyGui/_CommandFemConstraintSelfWeight.py
-        PyGui/_CommandFemElementFluid1D.py
-        PyGui/_CommandFemElementGeometry1D.py
-        PyGui/_CommandFemElementGeometry2D.py
-        PyGui/_CommandFemMaterialMechanicalNonlinear.py
-        PyGui/_CommandFemMaterialSolid.py
-        PyGui/_CommandFemMaterialFluid.py
-        PyGui/_CommandFemMesh2Mesh.py
-        PyGui/_CommandFemMeshBoundaryLayer.py
-        PyGui/_CommandFemMeshClear.py
-        PyGui/_CommandFemMeshGmshFromShape.py
-        PyGui/_CommandFemMeshGroup.py
-        PyGui/_CommandFemMeshNetgenFromShape.py
-        PyGui/_CommandFemMeshPrintInfo.py
-        PyGui/_CommandFemMeshRegion.py
-        PyGui/_CommandFemResultShow.py
-        PyGui/_CommandFemResultsPurge.py
-        PyGui/_CommandFemSolverCalculix.py
-        PyGui/_CommandFemSolverControl.py
-        PyGui/_CommandFemSolverRun.py
-        PyGui/_CommandFemSolverZ88.py
         PyGui/_TaskPanelFemElementFluid1D.py
         PyGui/_TaskPanelFemElementGeometry1D.py
         PyGui/_TaskPanelFemElementGeometry2D.py
